api-server/langchain_server/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain_google_genai
streamlit
langchain-core
firebase_admin
langserve
fastapi
gunicorn
sse_starlette
uvicorn
requests
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
torch
torchvision
torchaudio
transformers
openai
tiktoken
sentence-transformers
faiss-cpu
chromadb
pydantic
httpx
pytest
pytest-asyncio
black
isort
mypy
ruff
informatica
